Played around with erb rendering options
Settled on using '-' to indicate if there is a newline or not. The
rationale is that in a markup language like HTML the uglyness of the
source is not as important as Markdown (IMO). The API doc should be
readable as plain text... groan... going to have to fix those tables as
well now